Substantial update adding, in particular, "margposteriorsize" to compute the marginal posterior distribution of population size, and "discretemle" to compute the profile likelihood of the population size.
"margposteriorsize" uses a rejection sampler and appears to be algorithmically and computationally correct. However, the solutions are not credible, so I will need to think about it some more.

"discretemle" maximizes over a (given) set of population sizes using a mixture of optim() to get close in the continuous space of class sizes and then a sampler to search locally for the discrete set of class sizes that is the maximum. This seems to work well.

Both exploit parallalization using PVM/MPI via the "parallel=" argument. This speeds up computation considerable and works on mosix and student.

This commit os for the C side.

git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/rdsworkinggroup/software/size@102 adfe0e35-f9fe-4921-b9ec-9a706e04b580