1) Added code for the G. Andretta and G. M. Kaufman, "Estimation of finite population properties when sampling is without replacement and proportional to magnitude," J. Amer. Statist. Assoc., vol. 81, no. 395, pp. 657-666, 1986. This is basically Krista's thesie code.
2) Added code for a Polya urn scheme. The scheme samples from a non-parametricposterior for the full population (including the disease status) so that aposterior mean of the HIV status is a natural estimator to be compared to theHT based on the SPPS.

The model requires an estimate of the sample inclusion probabilities todrive it. It does not need to be perfect but the closer the better. Thebasic call is:sN <- ppspolya(s=s,N=N,y=y,wts=wts)# sN$y is a draw from the posterior of the infection status# for the full population (here 715 in size)## Compute the posterior meanest <- mean(sN$y)where s is the sample in order, N is the (assumed known) total populationsize, y is the disease status (of the sample) and wts are the estimates ofthe inverse inclusion probabilities (for the sample). So basically, it is analternative to the H-T estimator that has the advantage of coming withnatural measure of uncertainty (i.e., posterior variance and posterior 95%HPD intervals).In the example code I used the V-H wts (i.e., wts = 1/s). They are betterthan V-H but not as good as H-T with SPPS.

This need more work.

3) Updated the MLE functions (discretemle) to explicitly return the MLE etc and fixed a small number of blemish bugs.

git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/rdsworkinggroup/software/size@110 adfe0e35-f9fe-4921-b9ec-9a706e04b580