Made minor code improvements including internal documentation.
Added C code for estimation of inclusion probabilities for a given population. This is a C version of Krista's "getincl" R function. The C version is called "getinclC". The C version of "getest" (to compute the SPPS iestimator) is called "getestC".
It is not much faster but is a good code base for future C work.



git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/rdsworkinggroup/software/size@127 adfe0e35-f9fe-4921-b9ec-9a706e04b580