The new function "posteriornbinom" is just like "posteriorsize" except it uses the negative binomial degree distribution in place of the Poisson-lognormal.
Also added a prior on the population size. The previous was a uniform on the integers from 1 through "maxN". The new one is a negative binomial with mean "mean.prior.size" and standard deviation "sd.prior.size". Note that the data effectively truncates the prior below the sample size, but the priors are specified to allow for different sample sizes on the same population.


git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/rdsworkinggroup/software/size@156 adfe0e35-f9fe-4921-b9ec-9a706e04b580