Changed the prior structure on the population size, N. Thinking about priors on population side is problematic.
The original prior was uniform on (n,maxN) where n is the sample size and maxN was a user specified upper bound on N.  The likelihoods often asymptoted out to infinity (i.e. their is real "mass" on large values on N. In addition this causes computational problems (the MCMC sampler takes a long time).  To get this prior set the parameter "sd.prior.size=-1" in "posteriordisease".

To fix this we added a Negative Binomial model for N. But this has too thin tails for large N and requires accurate specification of the prior mean and s.d. to get it right [based on simulations that Cori and I did] To get this prior set the parameters "mean.prior.size=N, sd.prior.size=N" in "posteriordisease". Here N=maxN/2 is the best guess at the population size. 

To fix this we added the new sample proportion prior.  It is based on the idea that the sample size is not chosen separately from the population size but is often chosen in relation to it.  So a simple prior is a uniform prior on the sampling proportion (i.e. the proportion of the population in the sample). This is between 0 and 1. So the simplest prior is uniform on n/N. For simplicity I have used a continuous distribution even though N is discrete. This translates to a closed form for the prior on N which has infinite mean (and variance) so allows for very large Ns.

The prior implemented is more general by allowing a Beta(1,beta) distribution on n/N. I have derived a closed form to the resulting prior on N. It is parametrized by the median of the prior distribution on N (median.prior.size) - something that researchers might have an idea about.

Note: beta = -log(2)/log(1-n/median.prior.size)
And

Prior(N) = beta*n*(N-n)^(beta-1) / N^(beta+1)   for N > n

The distribution has tail behavior 1/(N^2).
Simple algebra shows the mode of the prior is at 0.5*n*(beta+1)
and the median is given by 

n / (1-(1/2)^(1/beta))

To use this prior just specify "median.prior.size". The value median.prior.size=2*n leads to the uniform on the sample proportion.

Preliminary tests of this prior suggest it leads to values of N which are too low, but allows for very big Ns will still producing valid posteriors (because of the 1/(N^2) decline and the fact that it is a proper prior for N).  Further tests will tell....

I think this ends the current phase of "size" development. We have a credible procedure to estimate N by using "posteriordisease" with "median.prior.size" set to the best idea of the median population size. I note that this produces posteriors for the population size, the disease prevalence and the disease count. 


git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/rdsworkinggroup/software/size@163 adfe0e35-f9fe-4921-b9ec-9a706e04b580