Updated the CMP-based posterior functions to compute a value for the maximum population size used in the computations (maxN). Previously it was passed, but now a value is computed from the prior (The 90% quantile prior) and this value is used if a value is not passed.
There is a new argument to "posteriordisease" called "priordistribution" with
values "cmp", "flat", and "nbinom" for Conway-Maxwell-Poisson, uniform and
Negative-Binomial. The default is "cmp".

The best way to specify the prior is via the hyperparameter "mode.prior.size"
which specifies the mode of the prior distribution on the population size.
You can alternatively specify the hyperparameter "median.prior.size"
which specifies the median of the prior distribution on the population size,
or "mode.prior.sample proportion" which specifies the mode of the prior
distribution on the proportion of the population size in the sample.


git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/rdsworkinggroup/software/size@165 adfe0e35-f9fe-4921-b9ec-9a706e04b580