Added a new output component to "posteriordisease" called "predictive.degree".
This is a vector of length the maximum degree (K) (by default K=2*max(sample degree)). 
The kth entry is the posterior predictive mean proportion of persons with degree k.
That is, it is the posterior predictive distribution for degrees from the population.

The idea is to compare this to the actual degree distribution proportions in the population (in cases where we know them). Thus this is a helpful posterior predictive check to the quality of the model we are using for the degree distribution (e.g., the default CMP).  

This mirrors the "predictive.degree" output from the "posteriorsize" function as is meant to match its functionality. That is, the "predictive.degree" from "posteriorsize" can also be used for posterior predictive checks.

Note: This commit just renames "ppos" to "predictive.degree".



git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/rdsworkinggroup/software/size@173 adfe0e35-f9fe-4921-b9ec-9a706e04b580