Added new output components to "posteriordisease" and "posteriorsize" called
"predictive.degree.count"

  This is a vector of length the maximum degree (K) (by default K=2*max(sample degree)). 
    The kth entry is the posterior predictive number of persons with degree k.
      That is, it is the posterior predictive distribution of the number with each degree in the population.


      Also added to "posteriordisease" new outputs called:

      "predictive.degree.count0"

        This is a vector of length the maximum degree (K) (by default K=2*max(sample degree)). 
  The kth entry is the posterior predictive number of non-diseased persons with degree k.
    That is, it is the posterior predictive distribution of the number of non-diseased with each degree in the population.

"predictive.degree.count1"

  This is a vector of length the maximum degree (K) (by default K=2*max(sample degree)). 
        The kth entry is the posterior predictive number of diseased persons with degree k.
  That is, it is the posterior predictive distribution of the number of diseased with each degree in the population.

The idea is to compare this to the actual degree distribution counts in the population (in cases where we know them). Thus this is a helpful posterior predictive check to the quality of the model we are using for the degree distribution (e.g., the default CMP).  

This mirrors the "predictive.degree" output but reflects the overall population size. That is, the two may not match because of an interaction between population size and degree distribution - something of interest to us.



git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/rdsworkinggroup/software/size@174 adfe0e35-f9fe-4921-b9ec-9a706e04b580