Increased the version number to 0.86.
Made sure that the data order is the same as the time order given in the "recruit.time" variable. If no "recruit.time" variable is given then no change takes place.

Swapped the "reflect.time" so that times are ordered with higher numbers later in time. This should not effect anything, but is more natural.