Changed "burnin" to "warmup" and fixed imputations
Changed "burnin" to "warmup" throughout, to be a slightly better term for stop. Note that "warmup" is a simple renaming of "burnin".

Fixed a bug in the impute.visibility function where it was returning incorrect imputations for missing network sizes when the recruitment time was used, and the data file was not in recruitment time order.